requirements.txt
fastapi== 0.116.1
uvicorn==0.35.0
PyJWT==2.10.1
python-jose==3.5.0
passlib==1.7.4
pymysql==1.0.2
python-multipart==0.0.20
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.12.0
pydantic-settings==2.10.1
pydantic==2.11.7
redis==6.2.0
python-dotenv==1.1.1
email-validator==2.2.0